Workaround: preserve IPv6 ScopeId on clone
When a service is advertised we store AddressRecord objects for it.
Because these are created by calling DnsObject.Clone(), they
serialize/deserialize the objects' network representation, which drops
the ScopeId.

This Id is necessary for determining which interface a given link-local IPv6
address is valid on.

This works around a bug in the Makaretu.Dns library, instead of fixing
it at the source.

The problem has been reported as
https://github.com/jdomnitz/net-dns/issues/63.
The fork we currently use (https://github.com/newky2k/net-mdns) is not
open for new issues.